Add Renderer.printText for lua for title screen

+++ Engine/script/bindings/core/events/luaevents_engine.h
@@ -7,7 +7,7 @@ class BindingCore_Events_Engine
 {
 public:
     static LuaEvent createInitEngineEvent(LuaEngine* engine);
-    static LuaEvent createLoopEvent(LuaEngine* engine, int ticks);
+    static LuaEvent createLoopEvent(LuaEngine* engine, float ticks);
 };
 
 #endif // BINDINGCORE_EVENTS_ENGINE_H
